Make PaymentSheet sheets use `StripeBottomSheetLayout`
