+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/PaymentOptionsActivity.kt
@@ -9,14 +9,16 @@ import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.graphics.Color
 import androidx.lifecycle.ViewModelProvider
-import com.stripe.android.common.ui.BottomSheet
-import com.stripe.android.common.ui.rememberBottomSheetState
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.stripe.android.paymentsheet.ui.BaseSheetActivity
 import com.stripe.android.paymentsheet.ui.PaymentSheetFlowType.Custom
 import com.stripe.android.paymentsheet.ui.PaymentSheetScreen
 import com.stripe.android.paymentsheet.utils.applicationIsTaskOwner
 import com.stripe.android.uicore.StripeTheme
+import com.stripe.android.uicore.elements.bottomsheet.StripeBottomSheetLayout
+import com.stripe.android.uicore.elements.bottomsheet.rememberStripeBottomSheetState
 import kotlinx.coroutines.flow.filterNotNull
 
 /**
@@ -51,11 +53,19 @@ internal class PaymentOptionsActivity : BaseSheetActivity<PaymentOptionResult>()
 
         setContent {
             StripeTheme {
+                val systemUiController = rememberSystemUiController()
                 val isProcessing by viewModel.processing.collectAsState()
 
-                val bottomSheetState = rememberBottomSheetState(
+                val bottomSheetState = rememberStripeBottomSheetState(
                     confirmValueChange = { !isProcessing },
                 )
+
+                LaunchedEffect(systemUiController) {
+                    systemUiController.setNavigationBarColor(
+                        color = Color.Transparent,
+                        darkIcons = false,
+                    )
+                }
 
                 LaunchedEffect(Unit) {
                     viewModel.paymentOptionResult.filterNotNull().collect { sheetResult ->
@@ -65,8 +75,14 @@ internal class PaymentOptionsActivity : BaseSheetActivity<PaymentOptionResult>()
                     }
                 }
 
-                BottomSheet(
+                StripeBottomSheetLayout(
                     state = bottomSheetState,
+                    onUpdateStatusBarColor = { color ->
+                        systemUiController.setStatusBarColor(
+                            color = color,
+                            darkIcons = false,
+                        )
+                    },
                     onDismissed = viewModel::onUserCancel,
                 ) {
                     PaymentSheetScreen(viewModel, type = Custom)
