+++ stripe-ui-core/src/main/java/com/stripe/android/uicore/elements/bottomsheet/StripeBottomSheetLayout.kt
@@ -1,116 +1,40 @@
 package com.stripe.android.uicore.elements.bottomsheet
 
 import androidx.annotation.RestrictTo
+import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.ime
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.MaterialTheme
+import androidx.compose.material.ModalBottomSheetDefaults
 import androidx.compose.material.ModalBottomSheetLayout
-import androidx.compose.material.ModalBottomSheetState
-import androidx.compose.material.ModalBottomSheetValue
-import androidx.compose.material.rememberModalBottomSheetState
+import androidx.compose.material.ModalBottomSheetValue.Expanded
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.State
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberUpdatedState
-import androidx.compose.runtime.snapshotFlow
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.platform.LocalTextInputService
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.text.input.TextInputService
 import androidx.compose.ui.unit.dp
+import com.stripe.android.uicore.elements.bottomsheet.StripeBottomSheetState.DismissalType
 import com.stripe.android.uicore.stripeShapes
-import kotlinx.coroutines.CancellationException
-import kotlinx.coroutines.flow.first
 
-internal const val BottomSheetContentTestTag = "BottomSheetContentTestTag"
-
-@OptIn(ExperimentalMaterialApi::class)
 @RestrictTo(RestrictTo.Scope.LIBRARY)
-class StripeBottomSheetState(
-    val modalBottomSheetState: ModalBottomSheetState,
-    val keyboardHandler: BottomSheetKeyboardHandler,
-    val sheetGesturesEnabled: Boolean,
-) {
-
-    private var dismissalType: DismissalType? = null
-
-    suspend fun show() {
-        repeatUntilSucceededOrLimit(10) {
-            // Showing the bottom sheet can be interrupted.
-            // We keep trying until it's fully displayed.
-            modalBottomSheetState.show()
-        }
-
-        // Ensure that isVisible is being updated correctly inside ModalBottomSheetState
-        snapshotFlow { modalBottomSheetState.isVisible }.first { isVisible -> isVisible }
-    }
-
-    suspend fun awaitDismissal(): DismissalType {
-        snapshotFlow { modalBottomSheetState.isVisible }.first { isVisible -> !isVisible }
-        return dismissalType ?: DismissalType.SwipedDownByUser
-    }
-
-    suspend fun hide() {
-        dismissalType = DismissalType.Programmatically
-        // We dismiss the keyboard before we dismiss the sheet. This looks cleaner and prevents
-        // a CancellationException.
-        keyboardHandler.dismiss()
-        if (modalBottomSheetState.isVisible) {
-            repeatUntilSucceededOrLimit(10) {
-                // Hiding the bottom sheet can be interrupted.
-                // We keep trying until it's fully hidden.
-                modalBottomSheetState.hide()
-            }
-        }
-    }
-
-    @RestrictTo(RestrictTo.Scope.LIBRARY)
-    enum class DismissalType {
-        Programmatically,
-        SwipedDownByUser,
-    }
-}
-
-@OptIn(ExperimentalMaterialApi::class)
-@RestrictTo(RestrictTo.Scope.LIBRARY)
-@Composable
-fun rememberStripeBottomSheetState(
-    initialValue: ModalBottomSheetValue = ModalBottomSheetValue.Hidden,
-    confirmValueChange: (ModalBottomSheetValue) -> Boolean = { true },
-): StripeBottomSheetState {
-    val modalBottomSheetState = rememberModalBottomSheetState(
-        initialValue = initialValue,
-        confirmValueChange = confirmValueChange,
-        skipHalfExpanded = true,
-        animationSpec = tween(),
-    )
-
-    val keyboardHandler = rememberBottomSheetKeyboardHandler()
-
-    return remember {
-        StripeBottomSheetState(
-            modalBottomSheetState = modalBottomSheetState,
-            keyboardHandler = keyboardHandler,
-            sheetGesturesEnabled = false,
-        )
-    }
-}
+const val BottomSheetContentTestTag = "BottomSheetContentTestTag"
 
 /**
  * Renders the provided [sheetContent] in a modal bottom sheet.
  *
  * @param state The [StripeBottomSheetState] that controls the visibility of the bottom sheet.
  * navigate to a specific screen.
+ * @param onUpdateStatusBarColor Called when the status bar color needs to be updated. This is based
+ * on the expansion state of the sheet.
  * @param onDismissed Called when the user dismisses the bottom sheet by swiping down. You should
  * inform your view model about this change.
+ * @param sheetContent The content to render in the sheet
  */
 @OptIn(ExperimentalMaterialApi::class)
 @RestrictTo(RestrictTo.Scope.LIBRARY)
@@ -118,14 +42,28 @@ fun rememberStripeBottomSheetState(
 fun StripeBottomSheetLayout(
     state: StripeBottomSheetState,
     modifier: Modifier = Modifier,
+    onUpdateStatusBarColor: (Color) -> Unit = {},
     onDismissed: () -> Unit,
     sheetContent: @Composable () -> Unit,
 ) {
+    val scrimColor = ModalBottomSheetDefaults.scrimColor
+    val isExpanded = state.modalBottomSheetState.targetValue == Expanded
+
+    val statusBarColorAlpha by animateFloatAsState(
+        targetValue = if (isExpanded) scrimColor.alpha else 0f,
+        animationSpec = tween(),
+        label = "StatusBarColorAlpha",
+    )
+
+    LaunchedEffect(statusBarColorAlpha) {
+        onUpdateStatusBarColor(scrimColor.copy(statusBarColorAlpha))
+    }
+
     LaunchedEffect(Unit) {
         state.show()
 
         val dismissalType = state.awaitDismissal()
-        if (dismissalType == StripeBottomSheetState.DismissalType.SwipedDownByUser) {
+        if (dismissalType == DismissalType.SwipedDownByUser) {
             onDismissed()
         }
     }
@@ -139,7 +77,7 @@ fun StripeBottomSheetLayout(
             topStart = MaterialTheme.stripeShapes.cornerRadius.dp,
             topEnd = MaterialTheme.stripeShapes.cornerRadius.dp,
         ),
-        sheetGesturesEnabled = state.sheetGesturesEnabled,
+        sheetGesturesEnabled = false,
         sheetElevation = 0.dp,
         sheetContent = {
             Box(modifier = Modifier.testTag(BottomSheetContentTestTag)) {
@@ -148,47 +86,4 @@ fun StripeBottomSheetLayout(
         },
         content = {},
     )
-}
-
-private suspend fun repeatUntilSucceededOrLimit(
-    limit: Int,
-    block: suspend () -> Unit
-) {
-    var counter = 0
-    while (counter < limit) {
-        try {
-            block()
-            break
-        } catch (ignored: CancellationException) {
-            counter += 1
-        }
-    }
-}
-
-@RestrictTo(RestrictTo.Scope.LIBRARY)
-class BottomSheetKeyboardHandler(
-    private val textInputService: TextInputService?,
-    private val isKeyboardVisible: State<Boolean>,
-) {
-
-    suspend fun dismiss() {
-        if (isKeyboardVisible.value) {
-            @Suppress("DEPRECATION")
-            textInputService?.hideSoftwareKeyboard()
-            awaitKeyboardDismissed()
-        }
-    }
-
-    private suspend fun awaitKeyboardDismissed() {
-        snapshotFlow { isKeyboardVisible.value }.first { !it }
-    }
-}
-
-@RestrictTo(RestrictTo.Scope.LIBRARY)
-@Composable
-fun rememberBottomSheetKeyboardHandler(): BottomSheetKeyboardHandler {
-    val imeHeight = WindowInsets.ime.getBottom(LocalDensity.current)
-    val isImeVisibleState = rememberUpdatedState(newValue = imeHeight > 0)
-    val textInputService = LocalTextInputService.current
-    return BottomSheetKeyboardHandler(textInputService, isImeVisibleState)
 }
